+++ cypress-shared/integration/b2b/2.12.2-sales_admin_scenarios.spec.js
@@ -11,52 +11,25 @@ import {
 } from '../../support/b2b/utils.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
 import {
-  productShouldNotbeAvailableTestCase,
   salesUserShouldImpersonateNonSalesUser,
-  userShouldNotImpersonateThisUser,
   verifySession,
   stopImpersonation,
 } from '../../support/b2b/common.js'
 import {
   searchQuote,
   createQuote,
-  verifyQuotesAndSavedCarts,
   // discountSliderShouldNotExist,
   updateQuote,
   // rejectQuote,
   // filterQuote,
   filterQuoteByStatus,
-  quoteShouldbeVisibleTestCase,
 } from '../../support/b2b/quotes.js'
+import { salesAdminQuotesAccess } from '../../support/b2b/impersonation_quote_access.js'
 
-function QuotesAccess(
-  { organizationName, quotes },
-  organizationB,
-  organizationBQuote
-) {
-  quoteShouldbeVisibleTestCase(
-    organizationName,
-    quotes.Buyer2.quotes1,
-    organizationName
-  )
-  quoteShouldbeVisibleTestCase(
-    organizationName,
-    organizationBQuote.OrganizationAdmin.quotes1,
-    organizationB
-  )
-}
-
-describe('Organization A - Cost Center A1 - Sales Admin Scenario', () => {
+describe('Organization A - Cost Center A1 - Sales Admin Impersonation Scenario', () => {
   loginViaCookies({ storeFrontCookie: false })
 
-  const {
-    product,
-    nonAvailableProduct,
-    costCenter1,
-    users,
-    quotes,
-    gmailCreds,
-  } = b2b.OrganizationA
+  const { product, costCenter1, users, quotes, gmailCreds } = b2b.OrganizationA
 
   const { organizationName: organizationB, quotes: organizationBQuote } =
     b2b.OrganizationB
@@ -65,14 +38,6 @@ describe('Organization A - Cost Center A1 - Sales Admin Scenario', () => {
 
   loginToStoreFront(users.SalesAdmin, roleObject.SalesAdmin.role, gmailCreds)
   verifySession(b2b.OrganizationA, costCenter1.name, roleObject.SalesAdmin.role)
-  productShouldNotbeAvailableTestCase(nonAvailableProduct)
-
-  // Impersonate users
-  userShouldNotImpersonateThisUser(
-    roleObject.SalesAdmin.role,
-    roleObject.SalesManager.role,
-    users.SalesManager
-  )
 
   salesUserShouldImpersonateNonSalesUser(
     roleObject.SalesAdmin.role,
@@ -80,24 +45,15 @@ describe('Organization A - Cost Center A1 - Sales Admin Scenario', () => {
     users.Approver1
   )
 
-  verifyQuotesAndSavedCarts()
-  QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
-  searchQuote(quotes.Buyer.quotes1)
-
-  // filterQuote(costCenter1.name, organizationB)
-
-  // discountSliderShouldNotExist(quotes.Buyer2.quotes3)
-
   updateQuote(quotes.Buyer.quotes1, { discount: '20' })
   updateQuote(quotes.Buyer.quotes2, { notes: 'Notes' })
   const price = '250.00'
   const quantity = '10'
 
   updateQuote(quotes.Buyer.quotes6, { price, quantity })
-  // rejectQuote(quotes.Buyer.quotes3, roleObject.SalesAdmin.role)
   filterQuoteByStatus(STATUSES.ready, STATUSES.declined)
 
-  QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
+  salesAdminQuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
 
   const quote = 'IMPERSONATE_QUOTE_1'