+++ cypress-shared/integration/b2b/2.13.2-sales_manager_scenarios.spec.js
@@ -6,57 +6,20 @@ import b2b from '../../support/b2b/constants.js'
 import {
   ROLE_ID_EMAIL_MAPPING as roleObject,
   ROLE_DROP_DOWN,
-  STATUSES,
 } from '../../support/b2b/utils.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
 import {
-  productShouldNotbeAvailableTestCase,
   salesUserShouldImpersonateNonSalesUser,
-  userShouldNotImpersonateThisUser,
   verifySession,
   stopImpersonation,
 } from '../../support/b2b/common.js'
-import {
-  createQuote,
-  searchQuote,
-  filterQuoteByStatus,
-  quoteShouldbeVisibleTestCase,
-  quoteShouldNotBeVisibleTestCase,
-} from '../../support/b2b/quotes.js'
+import { createQuote, searchQuote } from '../../support/b2b/quotes.js'
+import { salesManagerQuotesAccess } from '../../support/b2b/impersonation_quote_access.js'
 
-function QuotesAccess(
-  { organizationName, quotes },
-  organizationB,
-  organizationBQuote
-) {
-  quoteShouldbeVisibleTestCase(
-    organizationName,
-    quotes.OrganizationAdmin.quotes1,
-    organizationName
-  )
-  quoteShouldbeVisibleTestCase(
-    organizationName,
-    quotes.Buyer2.quotes1,
-    organizationName
-  )
-  quoteShouldNotBeVisibleTestCase(
-    organizationName,
-    organizationBQuote.OrganizationAdmin.quotes1,
-    organizationB
-  )
-}
-
-describe('Organization A - Cost Center A1 - Sales Manager Scenario', () => {
+describe('Organization A - Cost Center A1 - Sales Manager Impersonation Scenario', () => {
   loginViaCookies({ storeFrontCookie: false })
 
-  const {
-    nonAvailableProduct,
-    users,
-    product,
-    costCenter1,
-    quotes,
-    gmailCreds,
-  } = b2b.OrganizationA
+  const { users, product, costCenter1, gmailCreds } = b2b.OrganizationA
 
   const { organizationName: organizationB, quotes: organizationBQuote } =
     b2b.OrganizationB
@@ -71,12 +34,6 @@ describe('Organization A - Cost Center A1 - Sales Manager Scenario', () => {
     costCenter1.name,
     roleObject.SalesManager.role
   )
-  productShouldNotbeAvailableTestCase(nonAvailableProduct)
-  userShouldNotImpersonateThisUser(
-    roleObject.SalesManager.role,
-    roleObject.SalesRepresentative.role,
-    users.SalesRep
-  )
 
   const impersonatedRole = ROLE_DROP_DOWN.OrganizationAdmin
 
@@ -85,11 +42,9 @@ describe('Organization A - Cost Center A1 - Sales Manager Scenario', () => {
     impersonatedRole,
     users.OrganizationAdmin1
   )
-  searchQuote(quotes.OrganizationAdmin.quotes1)
-  filterQuoteByStatus(STATUSES.pending)
-  QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
-
   const quote = 'IMPERSONATE_QUOTE_2'
+
+  salesManagerQuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
 
   createQuote({
     product,