+++ cypress-shared/integration/fedex-shipping/2.11-validate-surcharge-percentage.spec.js
@@ -4,18 +4,14 @@
 import { updateRetry, loginViaCookies } from '../../support/common/support'
 import { appSetting } from '../../support/fedex-shipping/outputvalidation'
 import { data } from '../../fixtures/fedex-shipping-fixtures/shippingRatePayload.json'
-import {
-  filterByShippingMethod,
-  updateSLASettings,
-} from '../../support/fedex-shipping/common.js'
+import { updateSLASettings } from '../../support/fedex-shipping/common.js'
 import {
   loadCalculateShippingAPI,
   validateCalculateShipping,
 } from '../../support/fedex-shipping/api_testcase'
-import sla from '../../support/fedex-shipping/sla'
 
 const prefix = 'Shipping Optimize'
-let amount = ''
+let shippingMethods = []
 const surchargeFlatRate = 0
 const surchargePercent = 30
 
@@ -29,9 +25,7 @@ describe('Modify SLA - Validate Surcharge Percentage in checkout', () => {
   it(`${prefix} - Verify shipping price`, updateRetry(3), () => {
     loadCalculateShippingAPI(data).then((response) => {
       validateCalculateShipping(response)
-      const shippingMethods = filterByShippingMethod(response, sla)
-
-      amount = shippingMethods[0].price
+      shippingMethods = response.body
     })
   })
 
@@ -46,14 +40,23 @@ describe('Modify SLA - Validate Surcharge Percentage in checkout', () => {
       // eslint-disable-next-line jest/valid-expect-in-promise
       loadCalculateShippingAPI(data).then((response) => {
         validateCalculateShipping(response)
-        const shippingMethods = filterByShippingMethod(response, sla)
 
-        const calculatePercentage = (amount * surchargePercent) / 100
-        const calculateFlatRate = amount + surchargeFlatRate
+        shippingMethods.forEach((shippingMethod) => {
+          response.body.forEach((res) => {
+            // eslint-disable-next-line vtex/prefer-early-return
+            if (res.shippingMethod === shippingMethod.shippingMethod) {
+              const calculatePercentage =
+                (shippingMethod.price * surchargePercent) / 100
 
-        expect(shippingMethods[0].price.toFixed(2)).to.equal(
-          (calculatePercentage + calculateFlatRate).toFixed(2)
-        )
+              const calculateFlatRate = shippingMethod.price + surchargeFlatRate
+
+              // eslint-disable-next-line jest/no-conditional-expect
+              expect(parseFloat(res.price.toFixed(2))).to.equal(
+                parseFloat((calculatePercentage + calculateFlatRate).toFixed(2))
+              )
+            }
+          })
+        })
       })
     }
   )