+++ cypress-shared/support/quick-order/product.js
@@ -1,0 +1,26 @@
+export const products = [
+  {
+    SKU: 'AB120',
+    Quantity: 2,
+  },
+  {
+    SKU: 'AB121',
+    Quantity: 3,
+  },
+  {
+    SKU: 'AB122',
+    Quantity: 5,
+  },
+  {
+    SKU: 'AB123',
+    Quantity: 10,
+  },
+  {
+    SKU: 'AB124',
+    Quantity: 1,
+  },
+  {
+    SKU: 'AB125',
+    Quantity: 20,
+  },
+]