+++ node/cypress.js
@@ -95,6 +95,8 @@ exports.open = async () => {
       integrationFolder: `${CY_FOLDER}/integration`,
       supportFile: `${CY_FOLDER}/support`,
       fixturesFolder: `${CY_FOLDER}/fixtures`,
+      pluginsFile: `${CY_FOLDER}/plugins/index.js`,
+      downloadsFolder: `${CY_FOLDER}/downloads`,
     },
   }
 
@@ -119,6 +121,8 @@ exports.run = async (test, config, addOptions = {}) => {
       integrationFolder: SPEC_PATH,
       supportFile: `${SPEC_PATH.split(path.sep)[0]}/support`,
       fixturesFolder: `${CY_FOLDER}/fixtures`,
+      pluginsFile: `${CY_FOLDER}/plugins/index.js`,
+      downloadsFolder: `${CY_FOLDER}/downloads`,
     },
     spec: test.specs,
     headed: config.base.cypress.runHeaded,