+++ CHANGELOG.md
@@ -6,6 +6,58 @@
 - [ENGINEERS-1076] Detection of mixed path specs before start the test
 
 ### Changed
+- Validate $ in minicart
+- In organization request, add delay in typing for postalcode and street
+- For create organization request update hardRetries from 2 to 1
+
+## [2.2.1] - 2023-01-18
+
+### Changed
+- Improvement of shopper location & location availability
+
+### Changed
+
+- [ENGINEERS-1062]
+   - Once organization request been approved from cypress tests we create same organization request again and verify the approved toast message
+   (Because of one2many feature tests are failing - Disable the tests)
+   - Improvements
+
+### Removed
+
+- [ENGINEERS-1058] - Don't login with the user which we would be assigning to a b2b organization
+
+## [2.2.0] - 2023-01-09
+
+### Added
+- [ENGINEERS-1037] - Added affirm promo test
+
+## [2.1.0] - 2023-01-06
+
+### Removed
+- [ENGINEERS-1033] - Drop duplicate organization request testcase
+
+### Changed
+- Paypal cypress improvement
+
+### Changed
+- Verify Fedex Shipping prices code improvements
+
+## [2.0.1] - 2022-12-27
+
+### Fixed
+- Bump decode-uri-component to 0.2.2
+
+## [2.0.0] - 2022-12-27
+
+### Added
+- Added verify excel data testcase 
+
+## [1.0.4] - 2022-12-27
+
+### Fixed
+- [ENGINEERS-987] Fix fork message for PRN
+
+### Changed
 - [ENGINEERS-985] & [ENGINEERS-986] Affirm selectors improvement
 
 ### Changed
@@ -19,6 +71,8 @@
 ### Fixed
 - [ENGINEERS-975] Improve b2b checkout tests
 
+### Fixed
+- [ENGINEERS-975] Improve b2b checkout tests
 ### Fixed
 - [ENGINEERS-962] Proper identification when a PR is from Fork